+++ src/lib/jukebox.h
@@ -11,6 +11,10 @@
 #ifndef MYMPD_LIB_JUKEBOX_H
 #define MYMPD_LIB_JUKEBOX_H
 
+#include <stdbool.h>
+
+struct t_partition_state;
+
 /**
  * Jukebox state
  */
@@ -24,5 +28,8 @@ enum jukebox_modes {
 
 enum jukebox_modes jukebox_mode_parse(const char *str);
 const char *jukebox_mode_lookup(enum jukebox_modes mode);
+
+bool jukebox_file_save(struct t_partition_state *partition_state);
+bool jukebox_file_read(struct t_partition_state *partition_state);
 
 #endif