+++ src/mympd_api/mympd_api_handler.c
@@ -675,8 +675,8 @@ void mympd_api_handler(struct t_mympd_state *mympd_state, struct t_partition_sta
                 json_get_string(request->data, "$.params.expression", 0, NAME_LEN_MAX, &sds_buf1, vcb_issearchexpression_song, &parse_error) == true &&
                 json_get_fields(request->data, "$.params.fields", &tagcols, FIELDS_MAX, &parse_error) == true)
             {
-                response->data = mympd_api_jukebox_list(partition_state, mympd_state->stickerdb, response->data, request->cmd_id, request->id,
-                        uint_buf1, uint_buf2, sds_buf1, &tagcols);
+                response->data = mympd_api_jukebox_list(partition_state, mympd_state->stickerdb, &mympd_state->album_cache,
+                        response->data, request->cmd_id, request->id, uint_buf1, uint_buf2, sds_buf1, &tagcols);
             }
             break;
         }